:truck: move containers/Home back to page/index
