+++ apps/blog/src/pages/index.tsx
@@ -1,12 +1,33 @@
+import React, { FC } from "react";
 import { GetStaticProps } from "next";
 import matter from "gray-matter";
 import path from "path";
 import fs from "fs";
+import { Heading, List, ListItem } from "@chakra-ui/react";
 
 import { MDX_SOURCE_PATH } from "@/constants/mdx";
-import Home, { HomeProps } from "@/containers/Home";
+import PostCard from "@/components/blog/PostCard";
 
 const MAX_POSTS_PER_PAGE = 5;
+
+interface HomeProps {
+  postsMeta: FrontMatter.Meta[];
+}
+
+const Home: FC<HomeProps> = ({ postsMeta }) => (
+  <>
+    <Heading as="h1" mx="5">
+      Howardism
+    </Heading>
+    <List my="2">
+      {postsMeta.map((meta) => (
+        <ListItem key={meta.date}>
+          <PostCard title={meta.title} description={meta.description} date={meta.date} tags={[]} />
+        </ListItem>
+      ))}
+    </List>
+  </>
+);
 
 export const getStaticProps: GetStaticProps<HomeProps> = async () => {
   const postsMeta = fs
